feat(shared): add rng helpers and refactor sim-runner
