+++ packages/shared/src/rng.test.ts
@@ -1,6 +1,6 @@
 import { test } from 'node:test';
 import assert from 'node:assert/strict';
-import { XorShift32 } from './rng';
+import { XorShift32, mulberry32, gaussian } from './rng';
 
 test('XorShift32 generates a deterministic sequence', () => {
   const rng = new XorShift32(1);
@@ -21,4 +21,18 @@ test('XorShift32 float handles max int without returning 1', () => {
   const rng = new XorShift32(1584200935);
   const v = rng.float();
   assert.ok(v < 1);
+});
+
+test('mulberry32 generates a deterministic sequence', () => {
+  const rng = mulberry32(1);
+  assert.ok(Math.abs(rng() - 0.6270739405881613) < 1e-9);
+  assert.ok(Math.abs(rng() - 0.002735721180215478) < 1e-9);
+});
+
+test('gaussian draws using provided RNG', () => {
+  const vals = [0.1, 0.2];
+  let i = 0;
+  const rng = () => vals[i++];
+  const g = gaussian(rng);
+  assert.ok(Math.abs(g - 0.6631399714746835) < 1e-9);
 });
